:hammer: Explode ern-container-gen utils
