+++ ern-container-gen/src/clearReactPackagerCache.ts
@@ -1,0 +1,9 @@
+import { shell } from 'ern-core'
+
+// TODO : [WINDOWS SUPPORT]
+export function clearReactPackagerCache() {
+  const TMPDIR = process.env.TMPDIR
+  if (TMPDIR) {
+    shell.rm('-rf', `${TMPDIR}/react-*`)
+  }
+}
