+++ ern-container-gen/src/consts.ts
@@ -1,0 +1,1 @@
+export const containerMetadataFileName = 'container-metadata.json'
