+++ ern-container-gen/src/generators/ios/IosGenerator.ts
@@ -8,16 +8,14 @@ import {
   log,
   NativePlatform,
 } from 'ern-core'
-import {
-  bundleMiniApps,
-  generatePluginsMustacheViews,
-  copyRnpmAssets,
-  injectReactNativeVersionKeysInObject,
-  sortDependenciesByName,
-  populateApiImplMustacheView,
-  addElectrodeNativeMetadataFile,
-  prepareDirectories,
-} from '../../utils'
+import { bundleMiniApps } from '../../bundleMiniApps'
+import { generatePluginsMustacheViews } from '../../generatePluginsMustacheViews'
+import { copyRnpmAssets } from '../../copyRnpmAssets'
+import { injectReactNativeVersionKeysInObject } from '../../injectReactNativeVersionKeysInObject'
+import { sortDependenciesByName } from '../../sortDependenciesByName'
+import { populateApiImplMustacheView } from '../../populateApiImplMustacheView'
+import { addElectrodeNativeMetadataFile } from '../../addElectrodeNativeMetadataFile'
+import { prepareDirectories } from '../../prepareDirectories'
 import fs from 'fs'
 import path from 'path'
 import xcode from 'xcode-ern'
