+++ ern-container-gen/src/getContainerMetadataPath.ts
@@ -1,0 +1,6 @@
+import { containerMetadataFileName } from './consts'
+import path from 'path'
+
+export function getContainerMetadataPath(containerPath: string): string {
+  return path.join(containerPath, containerMetadataFileName)
+}
