+++ ern-container-gen/src/reactNativeBundleIos.ts
@@ -1,0 +1,28 @@
+import { BundlingResult, reactnative, shell } from 'ern-core'
+import fs from 'fs'
+import path from 'path'
+
+export async function reactNativeBundleIos(
+  outDir: string
+): Promise<BundlingResult> {
+  const miniAppOutPath = path.join(
+    outDir,
+    'ElectrodeContainer',
+    'Libraries',
+    'MiniApp'
+  )
+  const bundleOutput = path.join(miniAppOutPath, 'MiniApp.jsbundle')
+  const assetsDest = miniAppOutPath
+
+  if (!fs.existsSync(miniAppOutPath)) {
+    shell.mkdir('-p', miniAppOutPath)
+  }
+
+  return reactnative.bundle({
+    assetsDest,
+    bundleOutput,
+    dev: false,
+    entryFile: 'index.ios.js',
+    platform: 'ios',
+  })
+}
