+++ ern-container-gen/src/types/ContainerGenResult.ts
@@ -1,0 +1,8 @@
+import { BundlingResult } from 'ern-core'
+
+export interface ContainerGenResult {
+  /**
+   *  Metadata resulting from the bundling
+   */
+  bundlingResult: BundlingResult
+}
