+++ ern-container-gen/src/writeFile.ts
@@ -1,0 +1,13 @@
+import fs from 'fs'
+
+export async function writeFile(filename: string, data: any) {
+  return new Promise((resolve, reject) => {
+    fs.writeFile(filename, data, err => {
+      if (err) {
+        reject(err)
+      } else {
+        resolve()
+      }
+    })
+  })
+}
